requirements.txt
# ===================== AI RetinoCare Requirements =====================

# --- Core ML stack ---
numpy==1.26.4
pandas==2.3.3
scikit-learn==1.7.2
scipy==1.15.3

# --- TensorFlow for M2 Mac ---
tensorflow-macos==2.16.2
tensorflow-metal==1.2.0

# --- Image processing ---
pillow==12.0.0
opencv-python-headless==4.11.0.86
albumentations==2.0.8
albucore==0.0.24

# --- Model utilities ---
protobuf==4.25.8
h5py==3.15.1

# --- Visualization ---
matplotlib==3.10.7
seaborn==0.13.2

# --- Streamlit App ---
streamlit==1.52.1
altair==6.0.0
pydeck==0.9.1

# --- PDF Generation ---
fpdf2==2.8.5
reportlab==4.4.4

# --- Misc utilities ---
PyYAML==6.0.3
tqdm==4.67.1
joblib==1.5.2
requests==2.32.5
rich==14.2.0
